phase_diagram/requirements_phase.txt
# 相图应用依赖库
numpy
scipy
matplotlib
pandas
rdkit
pubchempy
python-ternary
streamlit
torch
transformers
huggingface_hub
